Setting a property as part of a mock

+++ src/Concise/Console/ResultPrinter/Utilities/ProgressBar.php
@@ -36,7 +36,7 @@ class ProgressBar
 
     protected function calculateTotal(array $parts)
     {
-        $this->total = array_sum($parts);
+        $this->total = (int) array_sum($parts);
     }
 
     protected function colorSpaces($length, $colorName)
@@ -48,6 +48,7 @@ class ProgressBar
 
     /**
      * @param string $currentProgressBar
+     * @return string
      */
     protected function fillToSize($currentProgressBar)
     {
